refactor(3a): move AlphaProjectConfigUpdater into the app module
First step of redrawing the framework/app boundary. AlphaProjectConfigUpdater is
SkyBlock-Nerds domain logic (it migrates the alpha/project forum config) that
lived in discord-framework but is only used by the app. Move it to
net.hypixel.nerdbot.app.config; no framework class depended on it, so nothing in
the framework breaks. Behaviour is unchanged.
